phase-6-monitoring-and-evaluation/requirements.txt
chromadb>=0.4.0
langchain>=0.3.0
langchain-together>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
chainlit>=1.0.0
pypdf>=4.0.0
tiktoken>=0.7.0
arize-phoenix>=8.0.0
openinference-instrumentation-langchain>=0.1.0
arize-phoenix-otel>=0.8.0
opentelemetry-api>=1.0.0
opentelemetry-sdk>=1.0.0
